+++ DotaHeroes/API/Events/EventArgs/Hero/HeroAttackingEventArgs.cs
@@ -20,7 +20,7 @@ namespace DotaHeroes.API.Events.EventArgs.Hero
 
         public Features.Hero Target { get; }
 
-        public int Damage { get; set; }
+        public decimal Damage { get; set; }
 
         public DamageType DamageType { get; set; }
 
@@ -34,7 +34,7 @@ namespace DotaHeroes.API.Events.EventArgs.Hero
         /// <param name="damage"><inheritdoc cref="Damage" /></param>
         /// <param name="damageType"><inheritdoc cref="DamageType" /></param>
         /// <param name="isAllowed"><inheritdoc cref="IsAllowed" /></param>
-        public HeroAttackingEventArgs(Features.Hero hero, Features.Hero target, int damage, DamageType damageType, bool isAllowed)
+        public HeroAttackingEventArgs(Features.Hero hero, Features.Hero target, decimal damage, DamageType damageType, bool isAllowed)
         {
             Hero = hero;
             Target = target;
