+++ DotaHeroes/API/Events/EventArgs/Hero/HeroHealingEventArgs.cs
@@ -17,7 +17,7 @@ namespace DotaHeroes.API.Events.EventArgs.Hero
 
         public Features.Hero Healer { get; }
 
-        public float Heal { get; set; }
+        public decimal Heal { get; set; }
 
         public bool IsAllowed { get; set; }
 
@@ -28,7 +28,7 @@ namespace DotaHeroes.API.Events.EventArgs.Hero
         /// <param name="healer"><inheritdoc cref="Healer" /></param>
         /// <param name="heal"><inheritdoc cref="Heal" /></param>
         /// <param name="isAllowed"><inheritdoc cref="IsAllowed" /></param>
-        public HeroHealingEventArgs(Features.Hero hero, Features.Hero healer, float heal, bool isAllowed)
+        public HeroHealingEventArgs(Features.Hero hero, Features.Hero healer, decimal heal, bool isAllowed)
         {
             Hero = hero;
             Healer = healer;
